fix: Refactor imports to root index.ts
