--- packages/offix-offline/src/conflicts/index.ts
@@ -1,9 +1,0 @@
-export * from "./strategies/strategies";
-export * from "./ConflictLink";
-export * from "./strategies/ConflictListener";
-export * from "./state/ObjectState";
-export * from "./strategies/ConflictResolutionStrategy";
-export * from "./state/VersionedState";
-export * from "./BaseLink";
-export * from "./ConflictProcessor";
-export * from "./strategies/ConflictResolutionData";
