+++ packages/offix-offline/src/index.ts
@@ -1,3 +1,30 @@
-export * from "./conflicts";
-export * from "./offline";
+// Top level api
+export * from "./offline/OfflineLink";
+export * from "./conflicts/ConflictLink";
+
 export * from "./utils/helpers";
+export * from "./offline/network/NetworkStatus";
+export * from "./offline/network/WebNetworkStatus";
+export * from "./offline/network/CordovaNetworkStatus";
+export * from "./offline/events/OfflineQueueListener";
+export * from "./offline/OfflineMutationsHandler";
+export * from "./offline/storage/OfflineStore";
+export * from "./offline/events/CompositeQueueListener";
+export * from "./offline/events/ListenerProvider";
+export * from "./offline/events/OfflineQueueListener";
+export * from "./offline/LocalDirectiveFilterLink";
+export * from "./offline/storage/PersistentStore";
+export * from "./offline/storage/OfflineStore";
+export * from "./offline/storage/defaultStorage";
+export * from "./offline/processors/IDProcessor";
+export * from "./offline/processors/IResultProcessor";
+
+export * from "./conflicts/strategies/strategies";
+export * from "./conflicts/ConflictLink";
+export * from "./conflicts/strategies/ConflictListener";
+export * from "./conflicts/state/ObjectState";
+export * from "./conflicts/strategies/ConflictResolutionStrategy";
+export * from "./conflicts/state/VersionedState";
+export * from "./conflicts/BaseLink";
+export * from "./conflicts/ConflictProcessor";
+export * from "./conflicts/strategies/ConflictResolutionData";
