--- packages/offix-offline/src/offline/index.ts
@@ -1,16 +1,0 @@
-export * from "./network/NetworkStatus";
-export * from "./network/WebNetworkStatus";
-export * from "./network/CordovaNetworkStatus";
-export * from "./events/OfflineQueueListener";
-export * from "./OfflineMutationsHandler";
-export * from "./storage/OfflineStore";
-export * from "./OfflineLink";
-export * from "./events/CompositeQueueListener";
-export * from "./events/ListenerProvider";
-export * from "./events/OfflineQueueListener";
-export * from "./LocalDirectiveFilterLink";
-export * from "./storage/PersistentStore";
-export * from "./storage/OfflineStore";
-export * from "./storage/defaultStorage";
-export * from "./processors/IDProcessor";
-export * from "./processors/IResultProcessor";
